(cks-droidian) script: introduce a kernel source configurator script for Droidian
This script will automatically configure a Linux kernel
to be compiled and packaged with the Droidian build tools.

It's still in early development.

This initial version is a prototype for searching a file
on a public github repo matching a pattern,
using curl and a github api url.

Once the file is found, it's downloaded using curl and
a github raw url.

Finally, the file found, which should be a subtree-initiator-script, is executed.

The repository name, url, and other information required
by the git/curl operations, are defined in vars and grouped
in functions to allow easy and clear modification.

Signed-off-by: berbascum <berbascum@ticv.cat>