(subtree-init-dkpt) script: Add initial script
The introduced script is designed to initialize a git branch
containing a droidian kernel packaging template,
in a Linux kernel source dir, as subtree.

This adds the required stuff to configure the kernel and package it for Droidian.

Template's testing repoitory:
https://github.com/berbascum/droidian-kernel-packaging-templates

The flag --template-branch=<branch-name> is required as script argument.

ADVICE: Some of the functionss used are not present in this script.

This script is automatically called by the main configurator:
https://github.com/berbascum/droidian-kernel-build-helper-scripts/blob/droidian/configure-kernel-source-for-droidian.sh

Currently is not designed to be executed manually as parent script.

Signed-off-by: berbascum <berbascum@ticv.cat>