(cks-droidian) script-consts: improve generic consts var name defs
The constants config script was defining the constants
with the generic variable names required by the used functions.
The variable names was renamed to include the subtree name.
Then, the generic variable names are defined in the exec
function from the main configuration script.

Some vars in the config constants script are exported
to make avaliable to the child initiator script.
This commit moves the exports from the constants script
to the exec function in the main script.

Rename function subtree_initiator_dkpt_set_vars to
subtree_initiator_dkpt_set_consts

Reorder constants in the constants config script.

Instroduce git_args in the constants script.

Signed-off-by: berbascum <berbascum@ticv.cat>