(subtree-init-dkcf) child: implement support as child
Add support for the droiidian kernel packaging configurator script.

Check for the main configurator script consts.
If not detected, load the local defaults.

Encapsulate conditionally the local functions abort and info
to do not load them when this script is called from the
main configurator script.

The main configurator script configures the droidian dir as symlink,
so we need to check it first.
When droidian exists in the kernel source's root dir and if a symlink,
abort the subtree clone to avoid converting the symlink to a regular
directory.

Signed-off-by: berbascum <berbascum@ticv.cat>